produtos/urls.py
from django.urls import include, path
from .views import LojaViewSet, ProdutoViewSet
from rest_framework.routers import DefaultRouter

router = DefaultRouter()
router.register(r'lojas', LojaViewSet)      
router.register(r'produtos', ProdutoViewSet)

urlpatterns = [
    path('', include(router.urls)),
]